Move TLLog views to tllogs
